requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.2
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
